% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_allocation.R
\name{cluster_allocation}
\alias{cluster_allocation}
\title{cluster allocation}
\usage{
cluster_allocation(model)
}
\arguments{
\item{model}{is the EFLDA fitted model}
}
\value{
the cluster allocation matrix
}
\description{
cluster allocation
}
