% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsed_gibbs.R
\name{collapsed_gibbs}
\alias{collapsed_gibbs}
\title{Collapsed Gibbs sampling}
\usage{
collapsed_gibbs(
  data_DTM,
  K,
  alpha = NULL,
  beta = NULL,
  tau = NULL,
  p = NULL,
  type = "LDA",
  thin = 1,
  niter = 5000,
  warmup = 0.5,
  seed = 42,
  z_init = NULL,
  verbose = T,
  all.post = F,
  data_output = T
)
}
\arguments{
\item{data_DTM}{data in format DTM}

\item{K}{number of topic}

\item{alpha}{parameter of the Dirichlet}

\item{beta}{parameter of the Dirichlet}

\item{tau}{parameter of the EFD changes mean of clusters}

\item{p}{probabilities vector}

\item{type}{LDA or EFD or EFD_Phi}

\item{thin}{thinning period}

\item{niter}{number of iteration}

\item{warmup}{percentage of the warmup}

\item{seed}{of the analysis}

\item{z_init}{initial values of the topic, it is a list with D elements and each has length Nd}

\item{verbose}{if you want print of the iteration}

\item{all.post}{True if you want all the niter samples from the posterior}

\item{data_output}{True if you want the training data in the output}
}
\value{
list of the results from the gibbs sampling
}
\description{
Collapsed Gibbs sampling
}
