% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_alternative.R
\name{Corpus_alternative}
\alias{Corpus_alternative}
\title{Corpus alternative}
\usage{
Corpus_alternative(corpus)
}
\arguments{
\item{corpus}{a list with: x$doc and x$topic_words}
}
\value{
a matrix with column Word, Doc and Topic
}
\description{
Corpus alternative
}
