% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_alternative_word.R
\name{Corpus_alternative_word}
\alias{Corpus_alternative_word}
\title{Corpus from raw document}
\usage{
Corpus_alternative_word(doc_raw, K)
}
\arguments{
\item{doc_raw}{the raw matrix of the documents}

\item{K}{number of topic}
}
\value{
list of corpus and vocabulary
}
\description{
Corpus from raw document
}
