% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_CTM.R
\name{corpus_CTM}
\alias{corpus_CTM}
\title{Corpus from Correlated Topic Model}
\usage{
corpus_CTM(D, V, K, mu, Sigma, beta, eps)
}
\arguments{
\item{D}{number of documents}

\item{V}{number of words in the vocabulary}

\item{K}{number of topics}

\item{mu}{mean of the multivariate Normal}

\item{Sigma}{variance of the multivariate Normal}

\item{beta}{parameter of the Dirichlet}

\item{eps}{vector of (non-negative) means for the Poisson}
}
\value{
a corpus
}
\description{
Corpus from Correlated Topic Model
}
