% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_EFD.R
\name{corpus_EFD}
\alias{corpus_EFD}
\title{Corpus from Extended Flexible Dirichlet}
\usage{
corpus_EFD(D, V, K, alpha, tau, p, beta, eps)
}
\arguments{
\item{D}{number of documents}

\item{V}{number of words in a vocabulary}

\item{K}{number of topics}

\item{alpha}{parameter of the Dirichlet}

\item{tau}{parameter that changes the mean of the clusters}

\item{p}{vector of probabilities}

\item{beta}{parameter of the Dirichlet}

\item{eps}{vector of (non-negative) means for the Poisson}
}
\value{
a corpus
}
\description{
Corpus from Extended Flexible Dirichlet
}
