% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_LDA.R
\name{corpus_LDA}
\alias{corpus_LDA}
\title{Corpus created by LDA}
\usage{
corpus_LDA(D, V, K, alpha, beta, eps)
}
\arguments{
\item{D}{number of documents}

\item{V}{number of words in the vocabulary}

\item{K}{number of topics}

\item{alpha}{parameter of the Dirichlet}

\item{beta}{parameter of the Dirichlet}

\item{eps}{vector of (non-negative) means for the Poisson}
}
\value{
a corpus
}
\description{
Corpus created by LDA
}
