% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFD_expectation.R
\name{E.EFD}
\alias{E.EFD}
\title{Expected value of Extended Flexible Dirichlet}
\usage{
E.EFD(a, p_vect, t)
}
\arguments{
\item{a}{paramater of the Dirichlet}

\item{p_vect}{vector of probabilities}

\item{t}{parameter that changes the mean of the clusters}
}
\value{
the expected value
}
\description{
Expected value of Extended Flexible Dirichlet
}
