% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFLDA.R
\docType{package}
\name{EFLDA-package}
\alias{EFLDA}
\alias{EFLDA-package}
\title{EFLDA: Extended Flexible Latent Dirichlet Allocation}
\description{
Extended Flexible Latent Dirichlet Allocation. A tool useful for topic modelling analyses. Positive correlation between topic allowed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AliceGiampino/EFLDA}
}

}
\author{
\strong{Maintainer}: Alice Giampino \email{a.giampino@campus.unimib.it}

Authors:
\itemize{
  \item Roberto Ascari \email{roberto.ascari@unimib.it}
}

}
\keyword{internal}
