% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{perplexity}
\alias{perplexity}
\title{perplexity}
\usage{
perplexity(model, newdata = NULL, posterior_mean = FALSE)
}
\arguments{
\item{model}{the fitted model}

\item{newdata}{test dataset}

\item{posterior_mean}{use posterior mean estimates}
}
\value{
the value of the perplexity
}
\description{
perplexity
}
