% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEFDir.R
\name{rEFDir}
\alias{rEFDir}
\title{Random realization from Extended Flexible Dirichlet}
\usage{
rEFDir(n, alpha, p, tau, label = FALSE)
}
\arguments{
\item{n}{number of words}

\item{alpha}{parameter of the Dirichlet}

\item{p}{vector of probabilities}

\item{tau}{parameter that changes the mean of the clusters}

\item{label}{if True report the topic?}
}
\value{
random realization from EFD
}
\description{
Random realization from Extended Flexible Dirichlet
}
