% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEFDir.R
\name{rEFDir}
\alias{rEFDir}
\title{Random realization from extended flexible Dirichlet}
\usage{
rEFDir(n, alpha, p, tau, label = FALSE)
}
\arguments{
\item{n}{the number of values to generate.}

\item{alpha}{a vector with positive elements.}

\item{p}{a vector with positive elements summing to 1.}

\item{tau}{a vector with positive elements.}

\item{label}{logical. If TRUE, an additional column pointing the component of the mixture generating the observation is returned.}
}
\value{
A matrix with \code{n} rows, where each row represents a sample from the EFD.
}
\description{
Random realization from extended flexible Dirichlet
}
\section{Functions}{
\itemize{
\item \code{rEFDir()}: function generates random values from the extended flexible Dirichlet (EFD) distribution.

}}
\references{
{
Ongaro, A., Migliorati, S., Ascari, R. (2020). A new mixture model on the simplex. Statistics and Computing, \bold{30}, 749-770, doi: 10.1007/s11222-019-09920-x.
}
}
