% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(data_DTM)
}
\arguments{
\item{data_DTM}{is a DTM matrix}
}
\value{
data in the format DTM: documents as rows and words as columns
}
\description{
transform_data
}
